[IMP] mrp_attachment_mgmt: Improve tests to allow inheritance in other addons that extend it. TT33400
mrp_attachment_mgmt 14.0.1.1.0
